split the link checker into themed modules for testing purposes
